Implement Resonant Update functionality across device backends
- Introduced `ResonantUpdateConfig` structure to configure resonant updates, enhancing operational capabilities.
- Added `Resonant` interface methods for forward and backward updates, integrating resonant network functionalities into the device architecture.
- Updated CPU, CUDA, and Metal backends to bind and implement resonant operations, ensuring consistency across different device implementations.
- Enhanced existing tests and added new ones to validate the functionality and performance of resonant updates, ensuring correctness across all backends.
- Updated `kernels.metallib` to include new resonant operations, improving the overall functionality of the Metal backend.

+++ device/metal/layernorm/stats_parity_darwin_test.go
@@ -43,10 +43,8 @@ func TestLayerNormMetalStatsParity(t *testing.T) {
 		row := input[rowIndex*cols : (rowIndex+1)*cols]
 		sum := cpulayernorm.SumFloat32Native(row)
 		wantMean := float32(float64(sum) / float64(cols))
-		variance := float32(
-			float64(cpulayernorm.LayerNormSquaredDiffSumNative(row, wantMean)) / float64(cols),
-		)
-		wantInvStdDev := float32(1.0 / math.Sqrt(float64(variance+1e-5)))
+		variance := float64(cpulayernorm.LayerNormSquaredDiffSumNative(row, wantMean)) / float64(cols)
+		wantInvStdDev := float32(1.0 / math.Sqrt(variance+1e-5))
 
 		gotMean := stats[rowIndex*2]
 		gotInvStdDev := stats[rowIndex*2+1]
